feat: task editing in dashboard
