+++ components/dashboard/DashboardTasksTableTabs.tsx
@@ -20,9 +20,6 @@ export function DashboardTasksTableTabs({ tasks }: { tasks: TTask[] }) {
   const searchParams = useSearchParams();
   const currentTab = searchParams.get("tab") || TabsConstants.All;
   const [showCreateTaskForm, setShowCreateTaskForm] = useState(false);
-  const [showEditTaskForm, setShowEditTaskForm] = useState(false);
-  const [selectedTask, setSelectedTask] = useState<TTask | null>(null);
-  console.log("selected task", selectedTask)
   const queryClient = useQueryClient();
 
   const createTaskMutation = useMutation({
@@ -37,29 +34,8 @@ export function DashboardTasksTableTabs({ tasks }: { tasks: TTask[] }) {
     },
   });
 
-  const updateTaskMutation = useMutation({
-    mutationFn: (task: TTask) => tasksApi.updateTask.fn(task),
-    onSuccess: () => {
-      void queryClient.invalidateQueries({ queryKey: tasksApi.getTasks.key });
-      toast.success("Task updated successfully");
-      setShowEditTaskForm(false);
-    },
-    onError: () => {
-      toast.error("Failed to update task");
-    },
-  });
-
   const handleCreateTask = (task: TaskFormData) => {
     createTaskMutation.mutate(task as TTask);
-  };
-
-  const handleEditTask = (task: TaskFormData) => {
-    updateTaskMutation.mutate(task as TTask);
-  };
-
-  const handleTaskClick = (task: TTask) => {
-    setSelectedTask(task);
-    setShowEditTaskForm(true);
   };
   
   const handleTabChange = (value: string) => {
@@ -94,14 +70,12 @@ export function DashboardTasksTableTabs({ tasks }: { tasks: TTask[] }) {
           <DashboardTasksTable 
             type={TabsConstants.All} 
             tasks={tasks} 
-            onTaskClick={handleTaskClick}
           />
         </TabsContent>
         <TabsContent value={TabsConstants.WatchList}>
           <DashboardTasksTable 
             type={TabsConstants.WatchList} 
             tasks={tasks} 
-            onTaskClick={handleTaskClick}
           />
         </TabsContent>
       </Tabs>
@@ -112,27 +86,6 @@ export function DashboardTasksTableTabs({ tasks }: { tasks: TTask[] }) {
         onClose={() => setShowCreateTaskForm(false)}
         onSubmit={handleCreateTask as (data: TaskFormData) => void}
         mode={FORM_MODE.CREATE}
-      />
-      
-      {/* Edit Task Form */}
-      <TodoForm 
-        key={selectedTask?.id || 'edit-form'} // Force remount when task changes
-        open={showEditTaskForm} 
-        onClose={() => {
-          setShowEditTaskForm(false);
-          setSelectedTask(null);
-        }}
-        onSubmit={handleEditTask as (data: TaskFormData) => void}
-        mode={FORM_MODE.EDIT}
-        initialData={selectedTask ? {
-          id: selectedTask.id,
-          title: selectedTask.title,
-          description: selectedTask.description || "",
-          dueDate: selectedTask.dueDate || "",
-          tags: selectedTask.tags || [],
-          taskId: selectedTask.taskId,
-          status: selectedTask.status,
-        } : undefined}
       />
     </div>
   );
